Refactor: Swipe 훅 분리
